fix: Combined resource processing and refactored cloudformation comments
