+++ src/cfn-mcp-server/awslabs/cfn_mcp_server/stack_analysis/cloudformation_utils.py
